Budget-Preparation-Assistance-MongoDB/requirements.txt
flask
google-generativeai
matplotlib
ollama
openpyxl
pandas
plotly
python-dotenv
pymongo